requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
passlib[bcrypt]>=1.7.4
